# services/adapter-a/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI completion service adapter A with model management and monitoring",
        "title": "AI Adapter A API",
        "contact": {},
        "version": "1.2.0"
    },
    "basePath": "/",
    "paths": {
        "/batch": {
            "post": {
                "description": "Submit multiple prompts for batch processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Batch completion",
                "parameters": [
                    {
                        "description": "Batch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "model": {
                                    "type": "string"
                                },
                                "options": {
                                    "type": "object"
                                },
                                "prompts": {
                                    "type": "array"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "batch_id": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/batch/{batchId}": {
            "get": {
                "description": "Get status and results of a batch processing job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Get batch status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "batchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "batch_id": {
                                    "type": "string"
                                },
                                "results": {
                                    "type": "array"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Cancel a running batch processing job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Cancel batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "batchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/benchmark": {
            "post": {
                "description": "Start a performance benchmark test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performance"
                ],
                "summary": "Run benchmark",
                "parameters": [
                    {
                        "description": "Benchmark configuration",
                        "name": "benchmark",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "concurrent_requests": {
                                    "type": "number"
                                },
                                "duration": {
                                    "type": "number"
                                },
                                "test_type": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "benchmark_id": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/benchmark/{benchmarkId}": {
            "get": {
                "description": "Get results of a performance benchmark test",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "performance"
                ],
                "summary": "Get benchmark results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Benchmark ID",
                        "name": "benchmarkId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "benchmark_id": {
                                    "type": "string"
                                },
                                "results": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/complete": {
            "post": {
                "description": "Generate AI text completion using adapter A",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "completion"
                ],
                "summary": "Generate text completion",
                "parameters": [
                    {
                        "description": "Completion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.completeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.completeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/model/config": {
            "put": {
                "description": "Update model configuration parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "Update model configuration",
                "parameters": [
                    {
                        "description": "Configuration updates",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "max_tokens": {
                                    "type": "number"
                                },
                                "temperature": {
                                    "type": "number"
                                },
                                "top_p": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.UsageInfo": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer",
                    "example": 25
                },
                "prompt_tokens": {
                    "type": "integer",
                    "example": 10
                },
                "total_tokens": {
                    "type": "integer",
                    "example": 35
                }
            }
        },
        "main.completeRequest": {
            "type": "object",
            "properties": {
                "max_tokens": {
                    "type": "integer",
                    "example": 100
                },
                "model": {
                    "type": "string",
                    "example": "adapter-a"
                },
                "prompt": {
                    "type": "string",
                    "example": "Hello world"
                },
                "temperature": {
                    "type": "number",
                    "example": 0.7
                },
                "top_p": {
                    "type": "number",
                    "example": 0.9
                }
            }
        },
        "main.completeResponse": {
            "type": "object",
            "properties": {
                "completion": {
                    "type": "string",
                    "example": "A: Hello world response"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-09-16T10:30:00Z"
                },
                "model": {
                    "type": "string",
                    "example": "adapter-a"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123"
                },
                "usage": {
                    "$ref": "#/definitions/main.UsageInfo"
                }
            }
        }
    }
}